Reformat parameter definition yaml files so that an arbitrary number of them can be chained to define all the simulation parameters.
the level 0 'params' are the NEST kernel parameters,
level 0 'children' are the network definition parameters and all have a recursive tree structure (dictionaries with params/children keys) except 'connections' (list)
	- neuron models
	- synapse models
	- layers : hierarchical definition of network layers
	- connection models : models for the different types of connections used in <connections>
	- connections : population to population connections.
	- populations: information on which populations (ie layer/neuron model tuple) are recorded and how.
NB: Because 'connections' doesn't have a recursive structure (so that we don't have to invent keys for all the individual connections), it cannot be chained. Therefore please don't define a 'connection' keys in more than one file.
NBB: This commit doesn't allow to initialize the network in NEST out of the box.