Test for value type when merging trees:
When horizontally merging node, the algorithm 'ChainTrees' all the
values of the equivalent nodes of the different trees if they are all
mappings, and returns only one of the values if not all the values
are mapping.
Fix: test for types of value and handle differently lists and mappings
from strings, floats and int so that empty lists and mappings are
dismissed but not strings, floats and ints.